+++ integrations/slack/events/comparison.py
@@ -6,14 +6,14 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from cls.timekit import ExtendedDatetime as ExtDt
-from cls.timekit import Format
 from integrations.protocols import ActionStatus, CommandType
 from libs.data import lookup, modify, search
 from libs.datamodels import ComparisonResults
 from libs.domain.score import GameResult
 from libs.types import StyleOptions
 from libs.utils import formatter, validator
+from libs.utils.timekit import ExtendedDatetime as ExtDt
+from libs.utils.timekit import Format
 
 if TYPE_CHECKING:
     from integrations.protocols import MessageParserProtocol
