+++ integrations/slack/functions.py
@@ -6,12 +6,12 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from cls.timekit import Delimiter, Format
-from cls.timekit import ExtendedDatetime as ExtDt
 from integrations.base.interface import FunctionsInterface
 from integrations.protocols import ActionStatus
 from libs.data import lookup
 from libs.utils import validator
+from libs.utils.timekit import Delimiter, Format
+from libs.utils.timekit import ExtendedDatetime as ExtDt
 
 if TYPE_CHECKING:
     from slack_sdk.web import SlackResponse
