+++ integrations/standard_io/functions.py
@@ -4,8 +4,8 @@ integrations/standard_io/functions.py
 
 from typing import TYPE_CHECKING
 
-from cls.timekit import ExtendedDatetime as ExtDt
 from integrations.base.interface import FunctionsInterface
+from libs.utils.timekit import ExtendedDatetime as ExtDt
 
 if TYPE_CHECKING:
     from integrations.protocols import MessageParserProtocol
