+++ libs/commands/graph/summary.py
@@ -10,12 +10,12 @@ import pandas as pd
 import plotly.express as px  # type: ignore
 
 import libs.global_value as g
-from cls.timekit import Format
 from libs.data import loader
 from libs.datamodels import GameInfo
 from libs.functions import compose, message
 from libs.types import StyleOptions
 from libs.utils import formatter, graphutil, textutil
+from libs.utils.timekit import Format
 
 if TYPE_CHECKING:
     from pathlib import Path
