+++ libs/data/modify.py
@@ -11,13 +11,13 @@ from contextlib import closing
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from cls.timekit import ExtendedDatetime as ExtDt
-from cls.timekit import Format
 from integrations.protocols import ActionStatus, MessageStatus
 from libs.data import lookup
 from libs.functions import message
 from libs.types import StyleOptions
 from libs.utils import dbutil, formatter
+from libs.utils.timekit import ExtendedDatetime as ExtDt
+from libs.utils.timekit import Format
 
 if TYPE_CHECKING:
     from integrations.protocols import MessageParserProtocol
