+++ libs/utils/dictutil.py
@@ -6,10 +6,10 @@ import logging
 from typing import TYPE_CHECKING, Any, cast
 
 import libs.global_value as g
-from cls.timekit import ExtendedDatetime as ExtDt
 from libs.data import lookup
 from libs.domain.command import CommandParser
 from libs.utils import formatter
+from libs.utils.timekit import ExtendedDatetime as ExtDt
 
 if TYPE_CHECKING:
     from cls.config import SubCommand
